Decompose AdminPage into per-tab sibling components
Split the 1,142-line AdminPage into a 167-line shell that owns the
company list, tab selection, and URL state, plus UsersTab, SitesTab, and
CompaniesTab siblings that each own their data fetching, dialogs, and
CRUD handlers.

Refs #48
